Adjust Glance/Swift backend configuration
As per [1] The way of configuring a Swift back-end for Glance
has changed. This change has been done in such a way that when
using Keystone v3 for authentication, the current configuraiton
method breaks.

This patch adjusts the way the configuration is done in order
to achieve the same result as before - a single working Swift
backend for Glance if the deployment opts to use Swift as a
backend.

[1] http://docs.openstack.org/developer/glance/configuring.html#configuring-swift-configuration-file

Change-Id: I02b69cf5985e8788513db58f7e8015f8135b9d58
Closes-Bug: #1561947