Add os-brick to glance_pip_packages
In order to use cinder as a glance back-end, the os-brick
python package is required.

While the inclusion of the package could be dynamically
included, we already have cinder listed as an additional
back-end by default and the cost of including this library
by default is minimal, so this opts to keep it simple and
just always include it.

The python package list is also sorted alphabetically for
easier package name finding in the future.

Change-Id: Id7e4638e1addf902e6027f6375c794fd56a5039d
Closes-Bug: #1634497