Detach glance swift_store_auth_insecure from Keystone
When not using local swift, backed by local keystone, we need the option
to set the swift_storage_auth_insecure to be different to the
keystone_internaluri_auth_insecure value.

This patch sets up it's own glance_swift_store_auth_insecure variable
which defaults to keystone_internaluri_auth_insecure, but can be
changed.

Change-Id: Ib9da0d1f94ea1cd46aa91a1fce0e97070f402693