Perform an atomic policy file change
The policy.json file is currently read continually by the
services and is not only read on service start. We therefore
cannot template directly to the file read by the service
(if the service is already running) because the new policies
may not be valid until the service restarts. This is
particularly important during a major upgrade. We therefore
only put the policy file in place after the service restart.

This patch also tidies up the handlers and some of the install
tasks to simplify them and reduce the tasks/code a little.

Conflicts:
>------handlers/main.yml
>------tasks/glance_init_common.yml

Combined backport of:
https://review.openstack.org/458861
https://review.openstack.org/467397
https://review.openstack.org/467574

Change-Id: I81de53d8ddc4b462b878b412e53a6de219b71f86
(cherry picked from commit cadb6bda617e11217063542ef44662e1997f8042)