Use command instead of debug for handlers
Due to the debug message plugin the handler restart
messages show at the end of the playbook execution
which is a little confusing. Using debug also
requires setting changed_when to true which is a
little extra bit of code which we do not have to
carry.

Instead we use the command module which is simple,
works and less wordy.

Change-Id: I7096ca81dd6e1126926c95f3c2b7437d0c9d452f
(cherry picked from commit aa4910a1cc964ca2f7825f3924a70c6b43e4e216)