Ensure that services restart in a particular order
Currently when multiple services share a host, the
restart order is random. This is due to an unordered
dict being used to facilitate the mapping of services
to their groups, names and other options.

This patch implements changes to the role to ensure
that services on the same host are restarted in the
correct order when the software/config changes.

Change-Id: I52fc66f861ce98cc8299c84edcfd5f18d74306b3
(cherry picked from commit 545e362b05264cce20b3fc025726d821a39e08c8)