Switch from run_once back to inventory scoping
As it turns out, run_once executes a task
once per batch when executed from a serialised
playbook instead of once per set of play hosts
as previously thought.

In this patch we implement a combination of
dynamic inclusion, inventory scoping and
play host scoping to achieve the required
goal of only running it once, even when the
playbook is executed using limits.

We also remove the unnecessary extra local
fact refresh.

Backport note:
ansible_play_hosts in Ansible 2.1.1 provides
the hosts for the play serial batch and not
the hosts for the play as a whole. It is
therefore not suitable, so we leave the
task scoping as it was.

Change-Id: Ib864bf2c0603e8281aa0ba3979ae020e90fd0adc
(cherry picked from commit 6e74637fcdc5084abc7c9dfcebbd4f98e043dde4)