Revert "Only sync glance database on major versions"
The current major upgrade detection compares
the same variable against itself. It also takes
place after the new venv tag has been placed
and the local facts are refreshed so there's
no good way to figure out the origin version.

Additionally, the version_compare filter only
works with semver version strings which is
unnecessarily strict.

Additionally, there does not appear to be any
evidence in the glance documentation that
the db sync action requires services to be
shut down or any other orchestration, so the
previously implemented action of doing a db
sync on every playbook execution should be
just fine. If there is nothing for it to do (which
will be the case for a minor upgrade) then it
will do nothing. If there is something for it to
do (which will be the case for a major upgrade)
then it will do what needs to be done.

As such, there appears to be no point in
implementing this extra set of tasks and
conditionals.

This reverts commit 647c4e33f1fcba771f23f7aabf0dc061b22f79ab.

Change-Id: I3b5afae3b7abf58de0b71ec2e67df2c38b6da5e3