Revert "Only sync glance database on major versions"
The current major upgrade detection compares
the same variable against itself. It also takes
place after the new venv tag has been placed
and the local facts are refreshed so there's
no good way to figure out the origin version.

Additionally, the version_compare filter only
works with semver version strings which is
unnecessarily strict.

Additionally, there does not appear to be any
evidence in the glance documentation that
the db sync action requires services to be
shut down or any other orchestration, so the
previously implemented action of doing a db
sync on every playbook execution should be
just fine. If there is nothing for it to do (which
will be the case for a minor upgrade) then it
will do nothing. If there is something for it to
do (which will be the case for a major upgrade)
then it will do what needs to be done.

As such, there appears to be no point in
implementing this extra set of tasks and
conditionals.

This reverts the https://review.openstack.org/475470
part of commit 8a5fb5df53a25d4c45e070f15ab036bc5617cb76.

Change-Id: Ia51c0ae1e2f81f662f5262c9f3d95027830a7d86