Implement uWSGI for glance-api
As part of the Pike goals we are moving api services to run as WSGI
apps. glance-api service is set up as a wsgi app, and this patch moves
it over.

Since this is just a drop in replacement for the existing eventlet
service operators and deployers should notice no difference.

Implements: blueprint goal-deploy-api-in-wsgi
Change-Id: Ie5fbf437031be01682534c466b3737d090a81c57