Disable glance-registry for the v2 API by default
When using the glance v2 API the glance-registry service is
optional, and the intention is to remove the glance-registry
service in the S cycle.

Prior to https://review.openstack.org/521814 the
glance-registry service was enabled, but the glance-api
service was not set to allow it to manage the data. It
therefore seems better to disable the registry for the
v2 API by default.

Tests are added to validate that disabling the v1 API works,
and enabling the v2 registry works.

Backport note:
The Queens version of this patch also disabled the v1 API
by default. That is not being done here as Pike is a stable
branch.

Change-Id: I4c27aa0ca5b649e4fa76cfd0f326d80f50074db1