Template api/registry enablement settings properly
If a deployer sets the boolean vars to 'yes', 'no' or
other equivalent booleans valid for ansible, the glance
services will currently use those in the .conf files
and it will break.

This patch ensures that we cast the output into the
.conf files properly so that regardless of the input,
the output is correct.

The tests are adjusted to deliberately use one of the
Ansible boolean values instead of a python boolean to
validate that this works.

Change-Id: I267b97da618bbda05140d2a0332798fc77db06a5
(cherry picked from commit 4e0c72cfe97259f41fcb76fba7c3fe0c326e44a7)