Fix systemd init template for program override
The systemd-init template was not looking at the program_override
variable within each service's dictionary.

This also fixes glance-api so that it's running under uWSGI when the v1
API is disabled. Creating images from a remote URL is exclusive to the
v1 API and does not work when glance-api is run under uWSGI.

The libxml2-dev package is required by uWSGI and has been added to the
distro package list.

Additional options have been added to the uWSGI configuration to better
support requests containing chunked data (image uploads).

Change-Id: I14f3162a4666d770beec9746469021466fa4d449