Set glance images dir owners each time
After a deployment of Queens, glance's images directory comes up with
ownership of `root:root` when it should be `glance:glance`. This
prevents an end user from uploading an image.

This patch removes the stat check of the images directory and always
ensures the directory is created. If the directory already exists,
the create task will ensure that the directory has its ownership
set appropriately.

Closes-Bug: 1742951
Change-Id: I286d6fbdfdec932ea031916f5460710a23d68374