Use a venv with previous ansible version for upgrades
When doing the role upgrade, we run every playbook with current's
branch ansible. This is normal for the infrastructure bits (which
are running on current branch), but not for the installation of
the previous branch's role, which should use previous branch's
environment.

This patch also renames test-upgrade-pre to test-upgrade, to
be in line with other roles.

Depends-On: I7ebb045885dd645b820de2b7f75b46c755c015f6
Change-Id: I3e7198d368ef1d7373bb036f60a75d1ddfa4b45c
Related-Bug: #1741471