Move uwsgi to install task file
When deploying from source, certain distros have their own way of
laying services and deploying them.  In this case, Glance uses
eventlet rather than UWSGI when deploying from binary.

This patch moves it into it the specific install file so that
it does not have to be done by binary installs.

Change-Id: If22beb98464b3b41112ffd54a66e8d7603bda2ae