Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: Iefe83a8b6002f423503c11f860a2f85f555a2f0a
Partial-Bug: #1764470
(cherry picked from commit 622fa8d5cfb5ea5d220a69333f7a0c74b41e22fd)