vars: Do not install build dependencies for distro installs
The glance_distro_packages contain packages which are build dependencies
to the 'PIP' ones so we don't need to install them when we are using the
Glance distribution packages. This also renames the '*glance_wsgi_bin'
to '*glance_uwsgi_bin' since it refers to the actual uwsgi package.
Finally, we add the missing 'install_method' conditional when recording
the local facts.

Implements: blueprint openstack-distribution-packages
Change-Id: I063a73ca7d50a0ad0bee57ccc42604802ab08220