Cleanup directory creation tasks
Using the realpath filter, we can idempotently create/modify
the log directory whether it is a symlink or not. This allows
us to consolidate two tasks into the common directory creation
task.

The extra task to create the images directory can be consolidated
into the same common directory creation task.

We also convert the items list to yaml dictionary format to
match the desired style according to the style guide.

Change-Id: I46e8b7c3be005681efc8a520318e7945c9c9a259