Add systemd tags to include role
There are times when a deployer will need to reconfigure parts of
an environment and having a general purpose tag to run said operation
will be important especicially should the deployer be needing to
reconfigure systemd unit files in a downtime event. This change adds
a general purpose systemd tag where include_role and systemd is found
which will assit operators with day2 operational tasks.

Change-Id: Id46c2aa56cd5498d022d29cf102dcbd2014c086d
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>